% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_edges.R
\name{apply_kernel}
\alias{apply_kernel}
\title{Apply a kernel to an image matrix}
\usage{
apply_kernel(img, kernel)
}
\arguments{
\item{img}{Numeric matrix}

\item{kernel}{Numeric square matrix}
}
\value{
Numeric matrix after convolution
}
\description{
Apply a kernel to an image matrix
}
