% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_edges.R
\name{detect_edges}
\alias{detect_edges}
\title{Detects edges in a numeric image matrix using a kernel}
\usage{
detect_edges(img, kernel = matrix(c(-1, -1, -1, -1, 8, -1, -1, -1, -1), 3, 3))
}
\arguments{
\item{img}{Numeric matrix representing the image}

\item{kernel}{Numeric square matrix representing the filter kernel. Defaults to Laplacian.}
}
\value{
Numeric matrix of normalized edge values (range 0-1)
}
\description{
Detects edges in a numeric image matrix using a kernel
}
\examples{
img <- matrix(runif(100), 10, 10)
kernel <- matrix(c(-1,-1,-1,-1,8,-1,-1,-1,-1), 3,3)
detect_edges(img, kernel)
}
