% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Says hello}
\usage{
hello(word = "world")
}
\arguments{
\item{word}{Input of the type character}
}
\value{
character.
}
\description{
'hello()' inserts the word hello before a word, that should be a character, that is inserted.
}
\examples{
hello('wolf')
}
