% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_edges.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize a matrix to range \code{0-1}}
\usage{
normalize_matrix(mat)
}
\arguments{
\item{mat}{Numeric matrix}
}
\value{
Numeric matrix normalized to \verb{[0,1]}
}
\description{
Normalize a matrix to range \code{0-1}
}
